+++ packages/dsw-document-worker/dsw/document_worker/model/__init__.py
@@ -1,4 +1,0 @@
-from .context import DocumentContext
-
-
-__all__ = ['DocumentContext']